requirements.txt
python-telegram-bot
requests
psutil
subprocess32